version_1.01-79 pr fixes part 1
