+++ app/src/FDroid/java/luci/sixsixsix/powerampache2/data/ErrorHandlerImpl.kt
@@ -142,6 +142,7 @@ class ErrorHandlerImpl @Inject constructor(
 
                     is ServerUrlNotInitializedException ->
                         "ServerUrlNotInitializedException $exceptionString"
+
                     is ScrobbleException -> {
                         readableMessage = ""
                         ""
