+++ app/src/main/java/luci/sixsixsix/powerampache2/presentation/screens/settings/ampache_preferences/components/CategoryText.kt
@@ -1,0 +1,47 @@
+/**
+ * Copyright (C) 2025  Antonio Tari
+ *
+ * This file is a part of Power Ampache 2
+ * Ampache Android client application
+ * @author Antonio Tari
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+package luci.sixsixsix.powerampache2.presentation.screens.settings.ampache_preferences.components
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.TextUnit
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun CategoryText(
+    category: String,
+    textSize: TextUnit = 16.sp,
+    colour: Color = MaterialTheme.colorScheme.tertiary
+) {
+    Text(category.uppercase(),
+        modifier = Modifier.padding(horizontal = 2.dp).padding(top = 16.dp, bottom = 4.dp),
+        fontWeight = FontWeight.Bold,
+        fontSize = textSize,
+        color = colour
+    )
+}
