+++ app/src/main/res/values-it/strings.xml
@@ -235,4 +235,17 @@
     <string name="bmac_btn_text">Buy Me A Coffee</string>
     <string name="about_support_title">Supporta il progetto</string>
     <string name="about_links_title">Links</string>
+
+    <string name="error_cannotEditPreference">Impossibile Modificare la Preferenza</string>
+    <string name="settings_ampachePreferences_title">Ampache Preferences</string>
+    <string name="settings_ampachePreferences_subtitle">Preferences for your Ampache Server</string>
+    <string name="ampachePreferences_updated_toast">Preference Updated</string>
+    <string name="ampachePreferences_userPreferences_title">User</string>
+    <string name="albums_sortOrder_name">Nome</string>
+    <string name="albums_sortOrder_year">Anno</string>
+    <string name="albums_sortOrder_artist">Artista</string>
+    <string name="albums_sortOrder_rating">Rating</string>
+    <string name="albums_sortOrder_averageRating">Average Rating</string>
+    <string name="ampachePreferences_screen_title">Ampache Preferences</string>
+    <string name="ampachePreferences_systemPreferences_title">System</string>
 </resources>
